Split component into several components and reworked tests.
